pyproject.toml
[project]
name = "qsynth"
version = "0.1.0"
dependencies = [
    "pandas",
    "faker>=18.4.0",
    "faker_airtravel",
    "faker_marketdata",
    "faker_vehicle",
    "pyyaml",
    "croniter>=1.3.14",
    "pyarrow",
    "pandavro",
    "fastavro>=1.5.1,<2.0.0",
    "numpy>=1.15.4,<2.0.0",
    "pydantic>=2.0.0",
    "pytest"
]
requires-python = ">= 3.9"

#[tool.setuptools.packages.find]
# All the following settings are optional:
#where = ["."]  # ["."] by default
#include = ["qsynth*"]  # ["*"] by default
#exclude = ["tests*"]  # empty by default
#namespaces = false  # true by default

[project.scripts]
cli-name="qsynth.main:exec_cli"